Add Python 3.12 and small MAP4 fix

+++ tests/fingerprints/map4.py
@@ -52,7 +52,8 @@ def test_map4_raw_hashes_fingerprint(smallest_smiles_list, smallest_mols_list):
     X_skfp = map4_fp.transform(smallest_smiles_list)
 
     X_map4 = np.stack(
-        [map4_fp._calculate_single_mol_fingerprint(mol) for mol in smallest_mols_list]
+        [map4_fp._calculate_single_mol_fingerprint(mol) for mol in smallest_mols_list],
+        dtype=np.int64,
     )
 
     assert np.array_equal(X_skfp, X_map4)
@@ -73,7 +74,8 @@ def test_map4_sparse_bit_fingerprint(smallest_smiles_list, smallest_mols_list):
     )
     X_map4 = np.mod(X_map4, map4_fp.fp_size)
     X_map4 = csr_array(
-        [(np.bincount(x, minlength=map4_fp.fp_size) > 0) for x in X_map4]
+        [(np.bincount(x, minlength=map4_fp.fp_size) > 0) for x in X_map4],
+        dtype=np.int64,
     )
 
     assert np.array_equal(X_skfp.data, X_map4.data)
